Revert "Add Wallet Context"
