--- frontend/src/config.js
@@ -1,1 +1,0 @@
-export const API_AUTH_BASE_URL = '';
