+++ frontend/src/contexts/AuthContext.js
@@ -8,7 +8,9 @@ const baseURL = API_AUTH_BASE_URL;
 
 const AuthContext = createContext();
 
-const AuthProvider = ({ children }) => {
+export default AuthContext;
+
+export const AuthProvider = ({ children }) => {
   const [authTokens, setAuthTokens] = useState(() =>
     localStorage.getItem('authTokens')
       ? JSON.parse(localStorage.getItem('authTokens'))
@@ -107,5 +109,3 @@ const AuthProvider = ({ children }) => {
     </AuthContext.Provider>
   );
 };
-
-export { AuthContext, AuthProvider };
