--- frontend/src/utils/customerProtected.js
@@ -1,10 +1,0 @@
-import { Navigate } from 'react-router-dom';
-import { useContext } from 'react';
-import { AuthContext } from '../contexts/AuthContext';
-
-const CustomerProtectedRoute = ({ children }) => {
-  let { user } = useContext(AuthContext);
-  return user ? children : <Navigate to="/retailer-login" />;
-};
-
-export default CustomerProtectedRoute;
