--- frontend/src/views/LandingPage.css
@@ -1,141 +1,0 @@
-video {
-  top: 0%;
-  object-fit: cover;
-  width: 100%;
-  height: 100vh;
-  position: fixed;
-  z-index: -1;
-}
-
-.landing-page {
-  height: 100vh;
-  width: 100%;
-  justify-content: center;
-  align-items: center;
-  box-shadow: inset 0 0 0 1000px rgba(0, 0, 0, 0.6);
-  object-fit: contain;
-}
-
-.hero-container {
-  height: 73vh;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  object-fit: contain;
-}
-
-.hero-container > h1 {
-  color: #fff;
-  font-size: 100px;
-  margin-top: 0px;
-}
-
-.hero-container > h2 {
-  color: #fff;
-  font-size: 100px;
-}
-
-.hero-container > p {
-  margin-top: 8px;
-  color: #fff;
-  font-size: 32px;
-  font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande',
-    'Lucida Sans', Arial, sans-serif;
-}
-
-.hero-btns {
-  margin-top: 32px;
-}
-
-.hero-btns .btn {
-  margin: 6px;
-}
-
-.fa-angle-right {
-  margin-left: 4px;
-}
-
-@media screen and (max-width: 960px) {
-  .hero-container > h1 {
-    font-size: 70px;
-    margin-top: 0px;
-  }
-
-  .hero-container > h2 {
-    font-size: 70px;
-    margin-top: 0px;
-  }
-}
-
-@media screen and (max-width: 768px) {
-  .hero-container > h1 {
-    font-size: 50px;
-    margin-top: 0px;
-  }
-
-  .hero-container > h2 {
-    font-size: 50px;
-    margin-top: 0px;
-  }
-
-  .hero-container > p {
-    font-size: 30px;
-  }
-
-  .btn-mobile {
-    display: block;
-    text-decoration: none;
-  }
-
-  .btn {
-    width: 100%;
-  }
-}
-
-@media screen and (max-width: 400px) {
-  .hero-container > h1 {
-    font-size: 43px;
-    margin-top: 0px;
-  }
-
-  .hero-container > h2 {
-    font-size: 43px;
-    margin-top: 0px;
-  }
-}
-
-@media screen and (max-width: 350px) {
-  .hero-container > h1 {
-    font-size: 38px;
-    margin-top: 0px;
-  }
-
-  .hero-container > h2 {
-    font-size: 38px;
-    margin-top: 0px;
-  }
-
-  .hero-container > p {
-    font-size: 25px;
-    margin-top: 0px;
-  }
-}
-
-@media screen and (max-width: 300px) {
-  .hero-container > h1 {
-    font-size: 35px;
-    margin-top: 0px;
-  }
-
-  .hero-container > h2 {
-    font-size: 35px;
-    margin-top: 0px;
-  }
-
-  .hero-container > p {
-    font-size: 20px;
-    margin-top: 0px;
-  }
-}
