--- frontend/src/views/retailer/retailerLogin.js
@@ -1,11 +1,0 @@
-function ReailerLogin() {
-  return (
-    <>
-      <div>
-        <h1>Retailer Login</h1>
-      </div>
-    </>
-  );
-}
-
-export default ReailerLogin;
